requirements.txt
asgiref==3.10.0
bleach==6.3.0
Django==5.2.8
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
git-filter-repo==2.47.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
webencodings==0.5.1
